fully move client config into client-sdk include directories
