+++ libs/client-sdk/include/launchdarkly/client_side/bindings/c/config/builder.h
@@ -3,8 +3,9 @@
 
 #pragma once
 
-#include <launchdarkly/bindings/c/config/client_side/config.h>
-#include <launchdarkly/bindings/c/config/common/logging_builder.h>
+#include <launchdarkly/client_side/bindings/c/config/config.h>
+
+#include <launchdarkly/bindings/c/config/logging_builder.h>
 #include <launchdarkly/bindings/c/export.h>
 #include <launchdarkly/bindings/c/status.h>
