+++ libs/client-sdk/tests/client_config_test.cpp
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
-#include <launchdarkly/bindings/c/config/client_side/builder.h>
-#include "launchdarkly/config/client.hpp"
+
+#include <launchdarkly/client_side/bindings/c/config/builder.h>
+#include <launchdarkly/config/client.hpp>
 
 TEST(ClientConfigBindings, ConfigBuilderNewFree) {
     LDClientConfigBuilder builder = LDClientConfigBuilder_New("sdk-123");
