+++ libs/common/include/launchdarkly/bindings/c/config/config.h
@@ -1,9 +1,9 @@
 /** @file
  * This header for the LaunchDarkly C++ Client-side SDK has moved!
  *
- * Please use 'launchdarkly/client_side/bindings/c/config/builder.h' instead.
+ * Please use 'launchdarkly/client_side/bindings/c/config/config.h' instead.
  *
- * The original C++ Client-side 1.0 release shipped builder.h at this location,
+ * The original C++ Client-side 1.0 release shipped config.h at this location,
  * but it was moved to the client_side subdirectory to accommodate the C++
  * Server-side SDK. */
 #pragma once
@@ -11,4 +11,4 @@
 #pragma message( \
     "LaunchDarkly Client-side C++ SDK: ACTION REQUIRED: This header has moved!")
 #pragma message( \
-    "LaunchDarkly Client-side C++ SDK: Please import 'launchdarkly/client_side/bindings/c/config/builder.h' instead")
+    "LaunchDarkly Client-side C++ SDK: Please import 'launchdarkly/client_side/bindings/c/config/config.h' instead")
